feat: remove graph.ts
