+++ package.json
@@ -14,8 +14,7 @@
     "clean": "rm -rf dist",
     "build:cjs": "tsc --p ./config/cjs/tsconfig.json",
     "build:esm": "tsc --p ./config/esm/tsconfig.json",
-    "build:graph": "tsc --p ./tsconfig.graph.json && cp -R public dist",
-    "build": "yarn build:cjs && yarn build:esm && yarn build:graph"
+    "build": "yarn build:cjs && yarn build:esm"
   },
   "dependencies": {
     "d3": "^7.9.0",
